+++ project_frontend/src/utils/index.js
@@ -2,5 +2,14 @@ import { movePath } from './movePath';
 import { isLogined } from './isLogined';
 import { DateString } from './dateString';
 import { PrivateRoutes, RegionRequiredRoutes } from './privateRoutes';
+import { fileExtensionValid, fileSizeValid } from './fileUploadValidHandler';
 
-export { movePath, isLogined, DateString, PrivateRoutes, RegionRequiredRoutes };
+export {
+	movePath,
+	isLogined,
+	DateString,
+	PrivateRoutes,
+	RegionRequiredRoutes,
+	fileExtensionValid,
+	fileSizeValid,
+};
